pipeline/requirements.txt
pandas
pyarrow
psycopg2-binary
sqlalchemy
pymupdf
boto3
voyageai
pgvector
numpy
python-dotenv
tqdm
rapidfuzz
